# [linker] move StripEmbeddedLibraries into the linker