… src/Xamarin.Android.Build.Tasks/Linker/MonoDroid.Tuner/RemoveLibraryResourceZip.cs
… src/Xamarin.Android.Build.Tasks/Tasks/StripEmbeddedLibraries.cs