Fixes: https://github.com/xamarin/xamarin-android/issues/1092

The `StripEmbeddedLibraries` MSBuild task can take 1-2 seconds, and it
mainly removes `__AndroidLibraryProjects__.zip` from assemblies.

If we moved this to happen during linking, it has various benefits:
- The linker already has every assembly opened and loaded.
- We know if the linker is going to `Skip`/`Delete` an assembly, so we
  can likewise skip it.
- The linker writes all the assemblies out at the end, so we don't
  have a second "write" step.

Changes to make this happen:
- Removed the `StripEmbeddedLibraries` MSBuild task and related
  targets
- Removed `$(_AndroidStripFlag)` from our targets, since it is no
  longer used
- Created a new `StripEmbeddedLibraries` linker step that runs late
  during linking
- Removed a `RemoveLibraryResourceZip` linker step, as it seemed to be
  duplicative.

I timed before and after with the Xamarin.Forms test project:

    .\bin\Debug\bin\xabuild .\tests\Xamarin.Forms-Performance-Integration\Droid\Xamarin.Forms.Performance.Integration.Droid.csproj /p:Configuration=Release /t:Clean
    .\bin\Debug\bin\xabuild .\tests\Xamarin.Forms-Performance-Integration\Droid\Xamarin.Forms.Performance.Integration.Droid.csproj /p:Configuration=Release /t:Build /bl

Before:

     1233 ms  StripEmbeddedLibraries                     1 calls
    14925 ms  LinkAssemblies                             1 calls

After:

    15437 ms  LinkAssemblies                             1 calls

As expected, `LinkAssemblies` will be slightly slower. But since
`StripEmbeddedLibraries` is not called at all, we have a net gain of
around 700ms.

Once this has been merged and working for `Release` builds, I plan to
do some further research to find out if running the new
`StripEmbeddedLibraries` linker step will help for `Debug` builds. It
could be a net performance improvement, if the time taken to remove
these files improves deployment and app startup times.