@jonathanpeppers 2018-09-07
@radekdoulik this is the first time I've looked at the linker, please let me know if I'm doing something dumb!

Here are the before/after binlogs: [StripEmbeddedLibraries.zip](https://github.com/xamarin/xamarin-android/files/2362342/StripEmbeddedLibraries.zip)

It seems like I haven't _broken_ what gets stripped, here is before:
```
         Stripped formsviewgroup.jar
         The stripped library is saved as obj\Release\90\linksrc\FormsViewGroup.dll
         Stripped arch-core-common.jar
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Arch.Core.Common.dll
         Stripped arch-lifecycle-common.jar
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Arch.Lifecycle.Common.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Arch.Lifecycle.Runtime.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.Animated.Vector.Drawable.dll
         Stripped support-annotations.jar
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.Annotations.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.Compat.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.Core.UI.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.Core.Utils.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.Design.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.Fragment.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.Media.Compat.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.Transition.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.v4.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.v7.AppCompat.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.v7.CardView.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.v7.MediaRouter.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.v7.Palette.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.v7.RecyclerView.dll
         Stripped __AndroidLibraryProjects__.zip
         The stripped library is saved as obj\Release\90\linksrc\Xamarin.Android.Support.Vector.Drawable.dll
```

Then after:
```
         Stripped formsviewgroup.jar from FormsViewGroup.dll
         Stripped arch-core-common.jar from Xamarin.Android.Arch.Core.Common.dll
         Stripped arch-lifecycle-common.jar from Xamarin.Android.Arch.Lifecycle.Common.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Arch.Lifecycle.Runtime.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.Animated.Vector.Drawable.dll
         Stripped support-annotations.jar from Xamarin.Android.Support.Annotations.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.Compat.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.Core.UI.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.Core.Utils.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.Design.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.Fragment.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.Media.Compat.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.Transition.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.v4.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.v7.AppCompat.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.v7.CardView.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.v7.MediaRouter.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.v7.Palette.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.v7.RecyclerView.dll
         Stripped __AndroidLibraryProjects__.zip from Xamarin.Android.Support.Vector.Drawable.dll
```