@jonathanpeppers 2018-09-11
Hmm, result of test run:
```
 09-10 21:49:35.108  3910  3928 I mono-stdout: Mono.Android_Tests : 7508.918 ms
 09-10 21:49:35.108  3910  3928 I mono-stdout: Tests run: 108, Passed: 107, Failed: 0, Skipped: 1, Inconclusive: 0
 09-10 21:49:35.125  3910  3928 I NUnitLite: NUnit automated tests completed
...
 Executing: /Users/builder/android-toolchain/sdk/platform-tools/adb   shell run-as Mono.Android_Tests cat "/data/user/0/Mono.Android_Tests/files/.__override__/TestResults.xml"
   writing stdout to file: /Users/builder/jenkins/workspace/xamarin-android-pr-builder/xamarin-android/tests/../bin/TestRelease/TestResult-Mono.Android_Tests.xml
   Command /Users/builder/android-toolchain/sdk/platform-tools/adb   shell run-as Mono.Android_Tests cat "/data/user/0/Mono.Android_Tests/files/.__override__/TestResults.xml" failed with exit code 1 
```

We have some weird build errors currently... So will look into this tomorrow.