@jonathanpeppers 2018-09-20
I think this one is actually green now?

The 3 test failures are that TLS thing we've been seeing: https://jenkins.mono-project.com/job/xamarin-android-pr-builder/4080/testReport/